feat: complete core features and error handling
- Add listtotalcount per list via sequential API calls
- Add email filter with whitespace trimming
- Add connect_timeout to Guzzle client for timeout handling
- Verify all required error cases: missing env, invalid credentials,
  empty list, API timeout
- Fix timeout type: float instead of int for sub-second values
- use safeLoad for dotenv to avoid rebuild on env changes
   - Switch from createImmutable/load to createUnsafeMutable/safeLoad
   - Docker env vars now picked up on restart without --build
   - .env file remains as local fallback only

+++ src/Api/SalesAutopilotClient.php
@@ -16,7 +16,7 @@ class SalesAutopilotClient
         private string $baseUrl,
         private string $username,
         private string $password,
-        private int $timeout = 10,
+        private float $timeout = 10.0,  // int -> float
     ) {
         $this->http = new Client([
             'base_uri' => rtrim($baseUrl, '/') . '/',
